extract builder-style API into standalone file to reduce clutter
